ocr-correction-viklofg-sweocr/pyproject.toml
[project]
name = "sparv-sbx-ocr-correction-viklofg-sweocr"
version = "0.3.0"
description = "A sparv plugin for computing suggested OCR improvements."
readme = "README.md"
authors = [
    { name = "Språkbanken Text", email = "sb-info@svenska.gu.se" },
    { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "parallel-corpus>=0.1.2",
    "sparv-pipeline>=5.2.0",
    "transformers>=4.45.2",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    # "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    # "uncomment if you test on these interpreters:",
    # "Programming Language :: Python :: Implementation :: IronPython",
    # "Programming Language :: Python :: Implementation :: Jython",
    # "Programming Language :: Python :: Implementation :: Stackless",
    "Topic :: Utilities",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."sparv.plugin"]
sbx_ocr_correction_viklofg_sweocr = "sbx_ocr_correction_viklofg_sweocr"

[project.urls]
Homepage = "https://spraakbanken.gu.se"
Repository = "https://github.com/spraakbanken/sparv-sbx-ocr-correction"
"Bug Tracker" = "https://github.com/spraakbanken/sparv-sbx-ocr-correction/labels/project%3Aocr-correction-viklofg-sweocr"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.build.targets.wheel]
packages = ["src/sbx_ocr_correction_viklofg_sweocr"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "syrupy>=4.7.2",
    "sparv-pipeline-testing @ git+https://github.com/spraakbanken/sparv-pipeline-testing.git@v0.1.7",
]
